refactor(sbom): extract sbom-scanner main logic into reusable RunServer package function
